requirements.txt
PyYAML==6.0.2
pytest==8.3.5
openpyxl==3.1.5
